Safari-609.1.11.1/LayoutTests/fast/events/ios/activating-button-should-not-scroll-page-expected.txt

Tests that activating a button (focusing it and pressing the spacebar) does not cause the page to scroll.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.scrollY is 0
PASS successfullyParsed is true

TEST COMPLETE
